fixed int calc in sparse loops

+++ Hadrons/Modules/MSolver/A2AVectors.hpp
@@ -484,9 +484,9 @@ void TStagSparseA2AVectors<FImpl, Pack>::execute(void)
                         if(zgp==zg || zgp==(zg+1)%ns){
                             site[2]=zl;
                             if(zshift[tglb]%2 || par().inc==1){
-                                sparseSite[2]=site[2]/par().inc;
+                                sparseSite[2]=int(site[2]/par().inc);
                             }else{
-                                sparseSite[2]=par().inc*site[2]/step + site[2]%2;
+                                sparseSite[2]=par().inc*int(site[2]/step) + site[2]%2;
                             }
                             for(int y=0;y<ns;y+=step){
                                 int yg=(yshift[tglb]+y)%ns;
@@ -495,9 +495,9 @@ void TStagSparseA2AVectors<FImpl, Pack>::execute(void)
                                     if(ygp==yg || ygp==(yg+1)%ns){
                                         site[1]=yl;
                                         if(yshift[tglb]%2 || par().inc==1){
-                                            sparseSite[1]=site[1]/par().inc;
+                                            sparseSite[1]=int(site[1]/par().inc);
                                         }else{
-                                            sparseSite[1]=par().inc*site[1]/step + site[1]%2;
+                                            sparseSite[1]=par().inc*int(site[1]/step) + site[1]%2;
                                         }
                                         for(int x=0;x<ns;x+=step){
                                             int xg=(xshift[tglb]+x)%ns;
@@ -506,9 +506,9 @@ void TStagSparseA2AVectors<FImpl, Pack>::execute(void)
                                                 if(xgp==xg || xgp==(xg+1)%ns){
                                                     site[0]=xl;
                                                     if(xshift[tglb]%2 || par().inc==1){
-                                                        sparseSite[0]=site[0]/par().inc;
+                                                        sparseSite[0]=int(site[0]/par().inc);
                                                     }else{
-                                                        sparseSite[0]=par().inc*site[0]/step + site[0]%2;
+                                                        sparseSite[0]=par().inc*int(site[0]/step) + site[0]%2;
                                                     }
                                                     for(int that=0;that<2;that++){
